inserting common into naive
